Task list is self contained and takes user as prop
